// src/api/schemas/bookmarkedTodoItemSchema.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keyo API
 * API for Keyo Polls app
 * OpenAPI spec version: 1.0.0
 */
import type { BookmarkedTodoItemSchemaTodoDueDate } from './bookmarkedTodoItemSchemaTodoDueDate';

/**
 * Schema for individual todo items in the list
 */
export interface BookmarkedTodoItemSchema {
  id: number;
  text: string;
  todo_completed: boolean;
  todo_due_date?: BookmarkedTodoItemSchemaTodoDueDate;
  created_at: string;
  updated_at: string;
}